+++ certbot/tests/cli_test.py
@@ -12,7 +12,7 @@ from acme import challenges
 
 import certbot.tests.util as test_util
 from certbot import cli
-from certbot import constants
+from certbot._internal import constants
 from certbot import errors
 from certbot.compat import os
 from certbot.compat import filesystem
