+++ certbot/tests/util.py
@@ -20,7 +20,7 @@ from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import serialization
 
 from certbot import configuration
-from certbot import constants
+from certbot._internal import constants
 from certbot import interfaces
 from certbot._internal import lock
 from certbot._internal import storage
