--- docs/api/constants.rst
@@ -1,9 +1,0 @@
-:mod:`certbot.constants`
------------------------------------
-
-.. automodule:: certbot.constants
-   :members:
-   :exclude-members: SSL_DHPARAMS_SRC
-
-.. autodata:: SSL_DHPARAMS_SRC
-   :annotation: = '/path/to/certbot/ssl-dhparams.pem'
